requirements.txt
pandas 
numpy
google-api-python-client
google-api-core
google-auth
google-auth-httplib2 
google-auth-oauthlib
oauth2client 
requests
yahoo_fin
datetime
bs4 